apps/api/pyproject.toml
[project]
name = "ieee-sponsorship-api"
version = "0.1.0"
description = "FastAPI backend for AI-powered sponsorship outreach"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "cryptography>=44,<46",
  "fastapi>=0.115,<1",
  "google-api-python-client>=2.170,<3",
  "google-auth>=2.40,<3",
  "google-auth-oauthlib>=1.2,<2",
  "httpx>=0.28,<1",
  "openai>=2.0.0,<3",
  "openpyxl>=3.1,<4",
  "orjson>=3.10,<4",
  "pydantic>=2.11,<3",
  "pydantic-settings>=2.6,<3",
  "python-dotenv>=1.1,<2",
  "python-multipart>=0.0.20,<1",
  "supabase>=2.13,<3",
  "uvicorn[standard]>=0.34,<1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "pytest-asyncio>=0.25,<1",
  "ruff>=0.11,<1"
]

[tool.uv]
package = false

[tool.pytest.ini_options]
pythonpath = ["app"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

